+++ tmf630-toolkit-paging-sorting-autoconfigure/src/main/java/org/opentmf/query/tmf630/config/Tmf630ExceptionHandlingAutoConfiguration.java
@@ -1,5 +1,6 @@
 package org.opentmf.query.tmf630.config;
 
+import org.opentmf.query.tmf630.advice.Tmf630FieldSelectionExceptionHandler;
 import org.opentmf.query.tmf630.advice.Tmf630PagingExceptionHandler;
 import org.opentmf.query.tmf630.advice.Tmf630RangeExceptionHandler;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
@@ -16,5 +17,10 @@ public class Tmf630ExceptionHandlingAutoConfiguration {
   @Bean
   public Tmf630PagingExceptionHandler tmf630PagingExceptionHandler() {
     return new Tmf630PagingExceptionHandler();
+  }
+
+  @Bean
+  public Tmf630FieldSelectionExceptionHandler tmf630FieldSelectionExceptionHandler() {
+    return new Tmf630FieldSelectionExceptionHandler();
   }
 }